feat: gifs, improvements
