+++ examples/shell/bash.gif
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cbb9f64d3aea47a6c3d61c532d0dfe2298653275b161f73571d447b539379231
+size 12805
