+++ examples/shell/fish.ascii
@@ -1,0 +1,54 @@
+> echo current shell: (basename (status fish-path))
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+────────────────────────────────────────────────────────────────────────────────
+> echo current shell: (basename (status fish-path))
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+────────────────────────────────────────────────────────────────────────────────
+> echo current shell: (basename (status fish-path))
+current shell: fish
+>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+────────────────────────────────────────────────────────────────────────────────
