+++ examples/shell/fish.gif
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9e463203567bc8971319bd763ba3e9eaaf92a00635208a30e50a2c56971d3b83
+size 41121
