+++ examples/shell/pwsh.ascii
@@ -1,0 +1,54 @@
+> echo "current shell: $ShellId
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+────────────────────────────────────────────────────────────────────────────────
+> echo "current shell: $ShellId"
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+────────────────────────────────────────────────────────────────────────────────
+> echo "current shell: $ShellId"
+current shell: Microsoft.PowerShell
+>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+────────────────────────────────────────────────────────────────────────────────
