+++ examples/shell/pwsh.gif
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a414378500e6786de8c7e318ed7938f7a4cd71acec63c6d6891ae94dec0dcb82
+size 24415
