+++ examples/shell/zsh.gif
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:73aba855095c5e75a257bef5b7fbe1ff647a58baf6b03b2666ae14b1ea3620d1
+size 12356
