+++ vhs.go
@@ -219,84 +219,8 @@ func (vhs *VHS) PauseRecording() {
 	vhs.recording = false
 }
 
-func (vhs *VHS) runShellCommand(format string, a ...interface{}) {
+func (vhs *VHS) runShellCommandf(format string, a ...interface{}) {
 	vhs.Page.MustElement("textarea").
 		MustInput(fmt.Sprintf(format, a...)).
 		MustType(input.Enter)
-}
-
-func getPrompt(vhs *VHS) Prompt {
-	switch vhs.Options.Shell {
-	case "", "bash":
-		return bash{}
-	case "zsh":
-		return zsh{}
-	case "fish":
-		return fish{}
-	case "pwsh":
-		return pwsh{}
-	default:
-		return generic{}
-	}
-}
-
-// Prompt defines how prompts set themselves up.
-type Prompt interface {
-	Setup(vhs *VHS)
-}
-
-var (
-	_ Prompt = bash{}
-	_ Prompt = zsh{}
-	_ Prompt = fish{}
-	_ Prompt = pwsh{}
-	_ Prompt = generic{}
-)
-
-type (
-	bash    struct{}
-	zsh     struct{}
-	fish    struct{}
-	pwsh    struct{}
-	generic struct{}
-)
-
-func (bash) Setup(vhs *VHS) {
-	prompt := vhs.Options.Prompt
-	if prompt == "" {
-		prompt = "\\[\\e[38;2;90;86;224m\\]> \\[\\e[0m\\]"
-	}
-	vhs.runShellCommand(` set +o history; unset PROMPT_COMMAND; export PS1="%s"; clear;`, prompt)
-}
-
-func (zsh) Setup(vhs *VHS) {
-	prompt := vhs.Options.Prompt
-	if prompt == "" {
-		prompt = `%F{blue bright dim}> %F{reset_color}`
-	}
-	vhs.runShellCommand(" clear; zsh --login --histnostore")
-	// PROMPT_SP: read about PROMPT_EOL_MARK
-	vhs.runShellCommand(` unsetopt PROMPT_SP; export PS1="%s"; clear`, prompt)
-}
-
-func (fish) Setup(vhs *VHS) {
-	prompt := vhs.Options.Prompt
-	if prompt == "" {
-		prompt = `function fish_prompt; echo -e "$(set_color --dim brblue)> $(set_color normal)"; end`
-	}
-	noGreeting := "function fish_greeting; end"
-	vhs.runShellCommand(` clear; fish --login --private -C '%s' -C '%s'`, noGreeting, prompt)
-}
-
-func (pwsh) Setup(vhs *VHS) {
-	prompt := vhs.Options.Prompt
-	if prompt == "" {
-		prompt = "Function prompt {Write-Host \"> \" -ForegroundColor Blue -NoNewLine; return \"`0\" }"
-	}
-	vhs.runShellCommand(` clear; pwsh -Login -NoLogo -NoExit -Command 'Set-PSReadLineOption -HistorySaveStyle SaveNothing; %s'`, prompt)
-}
-
-func (generic) Setup(vhs *VHS) {
-	// XXX: what should we do with prompt here?
-	vhs.runShellCommand(` clear; %s`, vhs.Options.Shell)
 }
